% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DumelleEtAl2021STLMM.R
\docType{package}
\name{DumelleEtAl2021STLMM-package}
\alias{DumelleEtAl2021STLMM}
\alias{DumelleEtAl2021STLMM-package}
\title{DumelleEtAl2021STLMM: Fit Spatio-Temporal Linear Mixed Models}
\description{
This package does stuff
}
\author{
\strong{Maintainer}: Michael Dumelle \email{Dumelle.Michael@epa.gov} (\href{https://orcid.org/0000-0002-3393-5529}{ORCID})

}
