% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLtoUTM.R
\name{LLtoUTM}
\alias{LLtoUTM}
\title{Transverse Mercator Projection}
\usage{
LLtoUTM(cm, lat, lon, xcol = "x", ycol = "y", minx = NULL, miny = NULL)
}
\arguments{
\item{cm}{iCentral meridian. A common choice is the
mean of the longitude values in your data set}

\item{lat}{Vector of latitudes.}

\item{lon}{Vector of longitudes.}

\item{xcol}{Name of the transformed x coordinates output by the function.}

\item{ycol}{Name of the transformed y coordinates output by the function.}

\item{minx}{A minimum x coordinate to center by. Defaults to setting the minimum
transformed coordinate to 0.}

\item{miny}{A minimum y coordinate to center by. Defaults to setting the minimum
transformed coordinate to 0.}
}
\value{
A list returning the transformed coordinates, central meridian argument,
minimum x coordinate, and minimum y coordinate.
}
\description{
Project latitude and longitude spatial coordinates using a Transverse Mercator
projection. Units are in kilometers, the minimum projected coordinates in
the x and y direction are 0, and projected coordinates are scaled to 1 kilometer.
}
