% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_covest_object.R
\name{make_covest_object}
\alias{make_covest_object}
\title{Title}
\usage{
make_covest_object(
  initial = NULL,
  estmethod,
  stcov,
  s_cor,
  t_cor,
  data_object,
  weights = NULL,
  chol = NULL,
  condition = NULL,
  max_options = NULL,
  optim_options = NULL,
  stempsv_options = NULL
)
}
\value{

}
\description{
Title
}
