% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_covparam_object.R
\name{make_covparam_object}
\alias{make_covparam_object}
\title{Title}
\usage{
make_covparam_object(
  s_de,
  s_ie,
  t_de,
  t_ie,
  st_de,
  st_ie,
  v_s,
  v_t,
  s_range,
  t_range,
  estmethod = NULL,
  stcov
)
}
\value{

}
\description{
Title
}
