% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_invert_object.R
\name{make_invert_object}
\alias{make_invert_object}
\title{Title}
\usage{
make_invert_object(
  covparam_object,
  chol,
  co = NULL,
  condition,
  h_s_large = NULL,
  h_t_large = NULL,
  h_s_small = NULL,
  h_t_small = NULL,
  logdet,
  m_index = NULL,
  o_index = NULL,
  s_cor,
  t_cor,
  xo,
  yo
)
}
\value{

}
\description{
Title
}
