% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_newdist_h.R
\name{make_newdata_h}
\alias{make_newdata_h}
\title{Title}
\usage{
make_newdata_h(
  newdata_coord1,
  data_coord1,
  newdata_coord2 = NULL,
  data_coord2 = NULL,
  distmetric = "euclidean"
)
}
\value{

}
\description{
Title
}
