% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stcovariance.R
\name{make_stcovariance}
\alias{make_stcovariance}
\alias{make_stcovariance.productsum}
\alias{make_stcovariance.sum_with_error}
\alias{make_stcovariance.product}
\title{Make a covariance matrix}
\usage{
make_stcovariance(covparam_object, h_s_large, h_t_large, s_cor, t_cor)

\method{make_stcovariance}{productsum}(covparam_object, h_s_large, h_t_large, s_cor, t_cor)

\method{make_stcovariance}{sum_with_error}(covparam_object, h_s_large, h_t_large, s_cor, t_cor)

\method{make_stcovariance}{product}(covparam_object, h_s_large, h_t_large, s_cor, t_cor)
}
\arguments{
\item{covparam_object}{A covparam object}

\item{h_s_large}{A spatial distance matrix of all spatio-temporal observations (if specified)}

\item{h_t_large}{A temporal distance matrix of all spatio-temopral observations (if specified)}

\item{s_cor}{The spatial correlation
\describe{
\item{\code{exponential}}{The exponential correlation (the default).}
\item{\code{spherical}}{The spherical correlation.}
\item{\code{gaussian}}{The Gaussian correlation.}
}}

\item{t_cor}{The temporal correlation
\describe{
\item{\code{exponential}}{The exponential correlation (the default).}
\item{\code{spherical}}{The spherical correlation.}
\item{\code{gaussian}}{The Gaussian correlation.}
\item{\code{tent}}{The tent (linear with sill) correlation.}
}}
}
\value{
A covariance matrix.
}
\description{
Make a covariance matrix
}
