% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.summary.stlmm}
\title{Print a Spatio-Temporal Linear Mixed Model Summary}
\usage{
\method{print}{summary.stlmm}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{A \code{stlmm} summary object.}

\item{digits}{The number of printing digits.}

\item{signif.stars}{The significance stars on p-values of the fixed effects.}

\item{...}{Additional arguments.}
}
\value{
A printed summary of the spatio-temporal linear mixed model fit.
}
\description{
Print a Spatio-Temporal Linear Mixed Model Summary
}
