% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.stlmm.R
\name{print.summary.stlmm}
\alias{print.summary.stlmm}
\title{Title}
\usage{
\method{print}{summary.stlmm}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\value{

}
\description{
Title
}
