% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.stlmm}
\title{Summarize a Spatio-Temporal Linear Mixed Model}
\usage{
\method{summary}{stlmm}(object, ...)
}
\arguments{
\item{object}{A \code{stlmm} object.}

\item{...}{Additional arguments.}
}
\value{
A list containing several objects
\describe{
\item{\code{Call}}{The original function call.}
\item{\code{FixedEffects}}{Fixed effects estimates and standard errors.}
\item{\code{CovarianceParameters}}{Covariance parameter estimates.}
\item{\code{CovarianceForms}}{The spatial, temopral, and spatio-temporal correlation forms.}
\item{\code{Residuals}}{Raw residuals.}
\item{\code{ObjectiveFn}}{Objective function value.}
}
}
\description{
Summarize a Spatio-Temporal Linear Mixed Model
}
