% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.stlmm}
\alias{predict.stlmm}
\title{Title}
\usage{
\method{predict}{stlmm}(
  object,
  newdata,
  interval = c("none", "confidence", "prediction"),
  se.fit = TRUE,
  predcov = FALSE,
  ...
)
}
\value{

}
\description{
Title
}
