% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stempsv.R
\name{stempsv}
\alias{stempsv}
\title{Title}
\usage{
stempsv(
  response,
  xcoord,
  ycoord = NULL,
  tcoord,
  h_options = NULL,
  h_response = NULL,
  h_s_large = NULL,
  h_t_large = NULL,
  stempsv_options = NULL
)
}
\value{

}
\description{
Title
}
